Process pipeline update;

+++ src/polyflexmd/data_analysis/data/constants.py
@@ -2,7 +2,7 @@ import numpy as np
 
 RAW_TRAJECTORY_DF_COLUMN_TYPES = {
     "t": np.uint64,  # From 0 to 18_446_744_073_709_551_615
-    "id": np.ushort,  # From 0 to 65_535
+    "id": np.uint32,  # From 0 to 4_294_967_295
     "type": np.ubyte,  # From 0 to 255
     "x": np.float32,
     "y": np.float32,
